Severe refactoring of the Gulp build setup, and a hopefully working shrinkwrapped NPM module setup. Relates to #4
