+++ gulpfile.js/clean.js
@@ -1,6 +1,11 @@
+/**
+ * @file Gulp tasks for cleaning temporary folders.
+ * @author Mark Lindhout <hetzblocker@marklindhout.com>
+ */
+
 const gulp = require('gulp')
 const del = require('del')
-const config = require('./gulpconfig.js')
+const config = require('../project.config.js')
 
 // Clean the build and dist folders
 gulp.task('clean', async function () {
