+++ gulpfile.js/extension-lint.js
@@ -1,6 +1,11 @@
+/**
+ * @file Gulp tasks for linting the extension using `web-ext`.
+ * @author Mark Lindhout <hetzblocker@marklindhout.com>
+ */
+
 const gulp = require('gulp')
 const webExt = require('web-ext').default
-const config = require('./gulpconfig.js')
+const config = require('../project.config.js')
 
 gulp.task('extension:lint:firefox', function () {
   return webExt.cmd.lint(config.webExtConfigFirefox, {
