+++ gulpfile.js/html.js
@@ -1,6 +1,11 @@
+/**
+ * @file Gulp tasks for compiling and copying HTML files.
+ * @author Mark Lindhout <hetzblocker@marklindhout.com>
+ */
+
 const gulp = require('gulp')
 const gulpTwig = require('gulp-twig')
-const config = require('./gulpconfig.js')
+const config = require('../project.config.js')
 
 gulp.task('html', function () {
   return gulp.src(config.sourceFolder + '/html/*.html.twig')
