+++ gulpfile.js/style.js
@@ -1,6 +1,11 @@
+/**
+ * @file Gulp tasks for compiling and copying style files.
+ * @author Mark Lindhout <hetzblocker@marklindhout.com>
+ */
+
 const gulp = require('gulp')
 const gulpLess = require('gulp-less')
-const config = require('./gulpconfig.js')
+const config = require('../project.config.js')
 
 gulp.task('style', function () {
   return gulp.src(config.sourceFolder + '/less/*.less')
