+++ package.json
@@ -27,7 +27,7 @@
     "gulp-twig": "^1.2.0",
     "gulp-uglify": "^3.0.2",
     "gulp-zip": "^4.2.0",
-    "jest-cli": "24.9.0",
+    "jest-cli": "^24.9.0",
     "jest-webextension-mock": "^3.6.0",
     "sharp": "^0.25.4",
     "standard": "^12.0.1",
